model_init bug fix
