+++ PyLCM/collision.py
@@ -6,6 +6,7 @@ from PyLCM.condensation import *
 from tqdm import tqdm
 import itertools
 
+ 
 def collection(dt, particles_list, rho_parcel, rho_liq, p_env, T_parcel, acc_ts, aut_ts):
     
     #shuffle the particle list for LSM (linear sampling method)
@@ -59,6 +60,7 @@ def collection(dt, particles_list, rho_parcel, rho_liq, p_env, T_parcel, acc_ts,
     
     return particles_list, acc_ts, aut_ts
 
+@jit(nopython=True) 
 def liquid_update_collection(particle1, particle2,acc_ts, aut_ts):
     
     # _int1: gains total individual mass
@@ -108,7 +110,7 @@ def liquid_update_collection(particle1, particle2,acc_ts, aut_ts):
         particle2 =  ptcl_int1
     
     return(particle1, particle2, acc_ts, aut_ts)
-
+@jit(nopython=True) 
 def same_weights_update(ptcl_int1, ptcl_int2, acc_ts, aut_ts):
     
     mass_crit = (seperation_radius_ts ** 3) * 4.0 / 3.0 * np.pi * rho_liq
@@ -139,7 +141,7 @@ def same_weights_update(ptcl_int1, ptcl_int2, acc_ts, aut_ts):
 
 import math
 import numpy as np
-
+@jit(nopython=True) 
 def determine_collision(dt, particle1, particle2, rho_parcel, rho_liq, p_env, T_parcel, half_length,nptcl):
     # Constants
     pi = math.pi
